#1 BFS: Set: Implementation is ready.
A note on the contains method's big O complexity: Buckets are backed by an array, which requires a linear scan. However, the elements are added at the first available position. By keeping track of the number of the total number of elements in the bucket, we can abandon the search the moment we checked all the added elements, regardless how big the available bucket is. For instance, if the bucket is empty or with a single matching element, it is an O(1) operation. If there are a lot of conflicts and the bucket is congested, linear scan should still be faster than using a linked list. The main disadvantage of this approach is that from time to time we need to resize the bucket when it reaches its capacity, but this can be mitigated by choosing the appropriate initial capacity.