#1 Dijkstra: Optimization: return early when end node is the next confirmed cheapest to reach node
By prioritizing cheapest to reach nodes, the algorithm guarantees that when a node is selected to be the next to calculate their neighbours cost, the node can't possibly be cheaper to reach that it is now. We can as well return at this point as there cannot be a path that would result in reaching this node cheaper.